+++ pkg/async/redis/common.go
@@ -1,6 +1,11 @@
 package redis
 
-import "fmt"
+import (
+	"fmt"
+
+	"github.com/Azure/open-service-broker-azure/pkg/async"
+	log "github.com/Sirupsen/logrus"
+)
 
 const (
 	workerSetName         = "workers"
@@ -15,4 +20,33 @@ func getActiveTaskQueueName(workerID string) string {
 
 func getWatchedTaskQueueName(workerID string) string {
 	return fmt.Sprintf("watched-tasks:%s", workerID)
+}
+
+func (e *engine) getTaskFromJSON(
+	taskJSON []byte,
+	queueName string,
+) (async.Task, error) {
+	task, err := async.NewTaskFromJSON(taskJSON)
+	if err != nil {
+		// If the JSON is invalid, remove the message from the queue, log this and
+		// move on. No other worker is going to be able to process this-- there's
+		// nothing we can do and there's no sense treating this as a fatal
+		// condition.
+		err := e.redisClient.LRem(queueName, -1, taskJSON).Err()
+		if err != nil {
+			return nil, fmt.Errorf(
+				`error removing malformed task from queue "%s"; task: %s: %s`,
+				queueName,
+				taskJSON,
+				err,
+			)
+		}
+		log.WithFields(log.Fields{
+			"queue":    queueName,
+			"taskJSON": taskJSON,
+			"error":    err,
+		}).Error("error decoding malformed task from queue")
+		return nil, nil
+	}
+	return task, nil
 }
